Uploaded progress thus far
The sconfig library (sconfig.h, sconfig.c) currently contains structures and programs intended to facilitate simulated geometric constructions (with digital "strings".) More specifically, the library (currently) supports numerical solvers that can be used to determine static string configurations with minimal total length, or constant length with a constant force applied to a single vertex. The solvers are relatively stable, but there are still a few wrinkles to iron out (especially in the constant length solver), and the algorithms are not yet optimized.  Other files (geo_graph, random_geo_graph, gen_configs) are primarily intended for testing and debugging, but may be useful when developing user interfaces.