Implemented prototype 'precision' solvers
Summary: I made substantial modifications to sc_minimizer_sd (to improve its utility for converging to singular configurations, and to make it more general-purpose), and added several new 'solver' functions combining it with the sc_minimizer_fin solver. Also: I made a first attempt at implementing a solver analogous to sc_minimizer_fin for the heuristic function (and an untested generalization of the 'precise' min. length solver to the const. length heuristic function.)